+++ src/app/admin/components/manage-order/manage-order.component.ts
@@ -29,12 +29,10 @@ export class ManageOrderComponent implements OnInit, CanComponentDeactivate {
   ) {}
 
   ngOnInit() {
-    this.route.data
-      .pipe(pluck('order'))
-      .subscribe((order: Order) => {
-        this.order = { ...order };
-        this.originalOrder = { ...order };
-      });
+    this.route.data.pipe(pluck('order')).subscribe((order: Order) => {
+      this.order = { ...order };
+      this.originalOrder = { ...order };
+    });
   }
 
   canDeactivate(): Promise<boolean> | boolean {
@@ -49,20 +47,18 @@ export class ManageOrderComponent implements OnInit, CanComponentDeactivate {
       return true;
     }
 
-    // Otherwise ask the user with the dialog service and return its
+    // Otherwise ask the user with the popup service and return its
     // promise which resolves to true or false when the user decides
     return this.popupService.confirm('Discard changes?');
   }
 
-   onSave() {
+  onSave() {
     const order = { ...this.order };
-    let savedOrder: Order;
-    savedOrder = this.orderService.editOrder(order);
-
-
-    this.originalOrder = { ...savedOrder };
-    this.order = { ...savedOrder };
-    this.onGoBack();
+    this.sub = this.orderService.editOrder(order).subscribe(o => {
+      this.originalOrder = { ...o };
+      this.order = { ...o };
+      this.onGoBack();
+    });
   }
 
   onGoBack() {