+++ src/app/admin/guards/order-resolve.guard.ts
@@ -1,30 +1,42 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, Router } from '@angular/router';
 
-import { Order } from '../../orders/models/order.model';
-import { OrderService } from '../../orders/services/order.service';
+import { Observable, of } from 'rxjs';
+import { catchError, map, take } from 'rxjs/operators';
+
+import { Order } from 'src/app/orders/models/order.model';
+import { OrderService } from 'src/app/orders/services/order.service';
 
 @Injectable()
 export class OrderResolveGuard implements Resolve<Order> {
   constructor(private orderService: OrderService, private router: Router) {}
 
-  resolve(route: ActivatedRouteSnapshot): Order {
+  resolve(route: ActivatedRouteSnapshot): Observable<Order> {
     if (!route.paramMap.has('orderID')) {
-      return {
+      return of({
         cartItems: [],
-        date: new Date(),
+        date: new Date().toISOString(),
         deliveryAddress: '',
         id: 0,
         name: ''
-      };
+      });
     }
 
     const id = +route.paramMap.get('orderID');
-    const order = this.orderService.getOrder(id);
-    if (!order) {
-      this.router.navigate(['/admin/orders']);
-    }
+    return this.orderService.getOrder(id).pipe(
+      map(order => {
+        if (order) {
+          return order;
+        }
 
-    return order;
+        this.router.navigate(['/admin/orders']);
+        return null;
+      }),
+      take(1),
+      catchError(() => {
+        this.router.navigate(['/admin/orders']);
+        return of(null);
+      })
+    );
   }
 }