+++ src/app/layout/header/header.component.ts
@@ -13,6 +13,8 @@ import { AutoUnsubscribe } from '../../core/decorators';
 })
 export class HeaderComponent implements OnInit {
   cartTotalCount: number;
+  expanded: boolean;
+
   private sub: Subscription = new Subscription();
 
   constructor(private cartService: CartService) {}
@@ -26,5 +28,9 @@ export class HeaderComponent implements OnInit {
     this.sub.add(this.cartService
       .getCount()
       .subscribe(count => (this.cartTotalCount = count)));
+  }
+
+  onToggleMenu() {
+    this.expanded = !this.expanded;
   }
 }